refactor: Consolidate FeistelCipher.Migration into FeistelCipher
BREAKING CHANGE: FeistelCipher.Migration module has been merged into FeistelCipher

- Merge all migration functions into main FeistelCipher module
- Remove lib/migration.ex file
- Update all references from FeistelCipher.Migration to FeistelCipher
- Simplify API: FeistelCipher.up() instead of FeistelCipher.Migration.up()
- All tests passing

Migration guide:
- Replace FeistelCipher.Migration.up() with FeistelCipher.up()
- Replace FeistelCipher.Migration.down() with FeistelCipher.down()
- Replace FeistelCipher.Migration.up_for_encryption() with FeistelCipher.up_for_encryption()
- Replace FeistelCipher.Migration.down_for_encryption() with FeistelCipher.down_for_encryption()
